refactor(project_split): separate plotter and collector process
This project was the original data collector and plotter for project
'Baronerd' which was renamed to project 'Air' within this work.
Along with the rename, the new project was split up into
multiple git repositories in order to separate development and
running environment concerns. This project is only concerned
with the data plotting portion of 'Air', leaving collecting of
the data and other functionality to the other repos within
this project.  The removing of 'baronerd' terms and other small
changes to the documentation takes place here.